Moving knex to be used by all the endpoints instead of the old
