+++ src/database.ts
@@ -1,88 +1,76 @@
 /* eslint no-console: 0 */
-import * as mysql from 'mysql2/promise';
-import { readdir } from 'fs/promises';
-import { join } from 'path';
+import { knex, Knex } from 'knex';
 
-let pool: mysql.Pool;
+let pool: Knex;
 
-async function getPool() {
+async function getPool(): Promise<Knex> {
 
   if (!pool) {
-    pool = mysql.createPool(await getSettings());
+    pool = knex(await getSettings());
   }
 
   return pool;
+}
+
+export async function getConnection(): Promise<Knex> {
+
+  return (await getPool());
 
 }
 
+export async function query(query: string, params: Knex.Value[] | Knex.ValueDict): Promise<Knex<any, any>> {
 
-export async function getConnection(): Promise<mysql.PoolConnection> {
-
-  return (await getPool()).getConnection();
-
-}
-
-export async function query(query: string, params?: any[]|string|Record<string,any>): Promise<[any, any]> {
-
-  return (await getPool()).query(query, params);
-
-}
-
-export async function checkPatches() {
-
-  // Get a list of known database patches.
-  const files = await readdir(join(__dirname, '..', 'mysql-schema'));
-  const patches = new Map<number, string>();
-  for(const file of files) {
-    const match = file.match(/^([0-9]+)-.*\.sql$/);
-    if (!match) continue;
-    patches.set(
-      parseInt(match[1]),
-      file
-    );
-  }
-
-  const result = await query('SELECT id FROM changelog');
-  for (const { id } of result[0]) {
-    if (!patches.has(id)) {
-      console.warn('Warning! Found an unknown database patch in the changelog. Patch id: ' + id);
-    }
-    patches.delete(id);
-  }
-
-  if (patches.size > 0) {
-    console.error('Error!! Database patches have not been applied. Please run the following .sql scripts:');
-    for (const fileName of patches.values()) {
-      console.error(`  - ${fileName}`);
-    }
-    throw new Error('Missing patches');
-  }
+  return (await getPool()).raw(query, params);
 
 }
 
 export default {
-  query,
   getConnection,
-  checkPatches,
+  query,
 };
 
 async function getSettings() {
 
-  let settings: any = {};
+  let connection: any = {};
+  let client: string = '';
+
+  // Declare explicitly the client to use, or try to infer it.
+  if (Object.keys(process.env).includes('PG_DATABASE')) {
+    client = 'pg';
+    connection = {
+      host: process.env.PG_HOST || '127.0.0.1',
+      port: process.env.PG_PORT || 5432,
+      user: process.env.PG_USER,
+      password: process.env.PG_PASSWORD,
+      database: process.env.PG_DATABASE,
+    };
+  } else if (Object.keys(process.env).includes('MYSQL_DATABASE')) {
+    client = 'mysql';
+    connection = {
+      host: process.env.MYSQL_HOST || '127.0.0.1',
+      port: process.env.MYSQL_PORT || 3306,
+      user: process.env.MYSQL_USER,
+      password: process.env.MYSQL_PASSWORD,
+      database: process.env.MYSQL_DATABASE,
+    }
+
+    if (process.env.MYSQL_INSTANCE_CONNECTION_NAME) {
+      connection.socketPath = '/cloudsql/' + process.env.MYSQL_INSTANCE_CONNECTION_NAME;
+    } else {
+      delete connection.host;
+      delete connection.port;
+    }
+  }
 
   // We are running in a local environment.
-  settings = {
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE,
+  return {
+    client,
+    connection,
+    searchPath: [
+      connection.user,
+      connection.database,
+      'public'
+    ],
+    pool: { min: 0, max: 10 },
   };
-
-  if (process.env.MYSQL_INSTANCE_CONNECTION_NAME) {
-    settings.socketPath = '/cloudsql/' + process.env.MYSQL_INSTANCE_CONNECTION_NAME;
-  } else {
-    settings.host = process.env.MYSQL_HOST;
-    settings.port = process.env.MYSQL_PORT || 3306;
-  }
-  return settings;
-
-}
+};
